LPS-193018 Abstract to base class for future reuse.
# Conflicts:
#	modules/apps/site/site-welcome-site-initializer/src/main/java/com/liferay/site/welcome/site/initializer/internal/instance/lifecycle/DefaultLayoutCreator.java
