+++ modules/apps/site/site-welcome-site-initializer/src/main/java/com/liferay/site/welcome/site/initializer/internal/instance/lifecycle/AddDefaultLayoutPortalInstanceLifecycleListener.java
@@ -5,10 +5,34 @@
 
 package com.liferay.site.welcome.site.initializer.internal.instance.lifecycle;
 
-import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
+import com.liferay.portal.instance.lifecycle.InitialRequestPortalInstanceLifecycleListener;
 import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
-import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.GroupConstants;
+import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutConstants;
+import com.liferay.portal.kernel.model.Role;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.model.role.RoleConstants;
+import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
+import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.security.permission.PermissionCheckerFactory;
+import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.LayoutLocalService;
+import com.liferay.portal.kernel.service.RoleLocalService;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.FriendlyURLNormalizer;
+import com.liferay.portal.util.PropsValues;
+import com.liferay.site.initializer.SiteInitializer;
 
+import java.util.List;
+
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -17,14 +41,100 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(service = PortalInstanceLifecycleListener.class)
 public class AddDefaultLayoutPortalInstanceLifecycleListener
-	extends BasePortalInstanceLifecycleListener {
+	extends InitialRequestPortalInstanceLifecycleListener {
+
+	@Activate
+	@Override
+	protected void activate(BundleContext bundleContext) {
+		super.activate(bundleContext);
+	}
 
 	@Override
-	public void portalInstanceRegistered(Company company) {
-		_defaultLayoutCreator.createDefaultLayout(company.getCompanyId());
+	protected void doPortalInstanceRegistered(long companyId) throws Exception {
+		Group group = _groupLocalService.getGroup(
+			companyId, GroupConstants.GUEST);
+
+		String friendlyURL = _friendlyURLNormalizer.normalizeWithEncoding(
+			PropsValues.DEFAULT_GUEST_PUBLIC_LAYOUT_FRIENDLY_URL);
+
+		Layout defaultLayout = _layoutLocalService.fetchLayoutByFriendlyURL(
+			group.getGroupId(), false, friendlyURL);
+
+		if (defaultLayout == null) {
+			defaultLayout = _layoutLocalService.fetchFirstLayout(
+				group.getGroupId(), false,
+				LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, false);
+
+			if (defaultLayout == null) {
+				String name = PrincipalThreadLocal.getName();
+
+				PermissionChecker permissionChecker =
+					PermissionThreadLocal.getPermissionChecker();
+
+				try {
+					User user = _getUser(companyId);
+
+					PrincipalThreadLocal.setName(user.getUserId());
+
+					PermissionThreadLocal.setPermissionChecker(
+						_defaultPermissionCheckerFactory.create(user));
+
+					ServiceContextThreadLocal.pushServiceContext(
+						new ServiceContext());
+
+					_siteInitializer.initialize(group.getGroupId());
+				}
+				finally {
+					PrincipalThreadLocal.setName(name);
+
+					PermissionThreadLocal.setPermissionChecker(
+						permissionChecker);
+
+					ServiceContextThreadLocal.popServiceContext();
+				}
+			}
+		}
+	}
+
+	private User _getUser(long companyId) throws PortalException {
+		Role role = _roleLocalService.fetchRole(
+			companyId, RoleConstants.ADMINISTRATOR);
+
+		if (role == null) {
+			return _userLocalService.getGuestUser(companyId);
+		}
+
+		List<User> adminUsers = _userLocalService.getRoleUsers(
+			role.getRoleId(), 0, 1);
+
+		if (adminUsers.isEmpty()) {
+			return _userLocalService.getGuestUser(companyId);
+		}
+
+		return adminUsers.get(0);
 	}
 
 	@Reference
-	private DefaultLayoutCreator _defaultLayoutCreator;
+	private PermissionCheckerFactory _defaultPermissionCheckerFactory;
+
+	@Reference
+	private FriendlyURLNormalizer _friendlyURLNormalizer;
+
+	@Reference
+	private GroupLocalService _groupLocalService;
+
+	@Reference
+	private LayoutLocalService _layoutLocalService;
+
+	@Reference
+	private RoleLocalService _roleLocalService;
+
+	@Reference(
+		target = "(site.initializer.key=com.liferay.site.initializer.welcome)"
+	)
+	private SiteInitializer _siteInitializer;
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
