+++ modules/apps/site/source-formatter-suppressions.xml
@@ -2,7 +2,6 @@
 
 <suppressions>
 	<checkstyle>
-		<suppress checks="CompanyIterationCheck" files="site-welcome-site-initializer/src/main/java/com/liferay/site/welcome/site/initializer/internal/instance/lifecycle/DefaultLayoutCreator\.java" />
 		<suppress checks="PersistenceCallCheck" files="site-memberships-test/src/testIntegration/java/com/liferay/site/memberships/service/persistence/test/UserGroupGroupRoleFinderTest\.java" />
 		<suppress checks="PersistenceCallCheck" files="site-test/src/testIntegration/java/com/liferay/site/service/persistence/test/GroupFinderTest\.java" />
 	</checkstyle>
